[AEA-4929] resolve review comments
